# openapi/wms-api.yaml
openapi: 3.1.0
info:
  title: WMS ShowCase API
  description: 倉庫管理システム（WMS）ShowCase のREST API定義
  version: 0.1.0
  contact:
    name: WMS ShowCase Project
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://{environment}.azurecontainerapps.io
    description: Azure Container Apps
    variables:
      environment:
        default: wms-showcase
...
paths:
...
  /api/v1/master/buildings:
    get:
      operationId: listBuildings
      summary: 棟一覧取得
      description: 指定倉庫に属する棟の一覧を取得する
      tags: [master-building]
      parameters:
        - name: warehouseId
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
          description: 対象の倉庫ID
        - name: buildingCode
          in: query
          schema:
            type: string
            maxLength: 10
          description: 棟コード（前方一致）
        - name: isActive
          in: query
          schema:
            type: boolean
          description: 有効/無効フィルタ
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: 棟一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingPageResponse'
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 指定倉庫が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createBuilding
      summary: 棟登録
      description: 棟マスタに新規棟を登録する
      tags: [master-building]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingRequest'
      responses:
        '201':
          description: 棟登録成功
          headers:
            Location:
              description: 作成されたリソースのURL
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 指定倉庫が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 棟コード重複
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                duplicateCode:
                  value:
                    errorCode: DUPLICATE_CODE
                    message: 同一倉庫内に同じ棟コードが既に存在します
  /api/v1/master/buildings/{id}:
    get:
      operationId: getBuilding
      summary: 棟取得
      description: 指定IDの棟マスタを1件取得する
      tags: [master-building]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      responses:
        '200':
          description: 棟取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 棟が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  value:
                    errorCode: BUILDING_NOT_FOUND
                    message: 指定された棟が見つかりません
    put:
      operationId: updateBuilding
      summary: 棟更新
      description: 指定IDの棟マスタを更新する。棟コード・倉庫IDは変更不可
      tags: [master-building]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBuildingRequest'
      responses:
        '200':
          description: 棟更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingDetail'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 棟が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
  /api/v1/master/buildings/{id}/toggle-active:
    patch:
      operationId: toggleBuildingActive
      summary: 棟無効化／有効化
      description: 指定IDの棟の有効/無効を切り替える。配下にエリアが存在する棟は無効化不可
      tags: [master-building]
      parameters:
        - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToggleActiveRequest'
      responses:
        '200':
          description: 無効化/有効化成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingToggleResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: 棟が存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
        '422':
          description: 配下にエリアが存在するため無効化不可
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                hasChildren:
                  value:
                    errorCode: CANNOT_DEACTIVATE_HAS_CHILDREN
                    message: 配下にエリアが存在するため無効化できません

  # --- Areas ---
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWTアクセストークン（httpOnly Cookie）

  # ============================================================
  # Schemas
  # ============================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
          description: エラーコード
          example: VALIDATION_ERROR
        message:
          type: string
          description: エラーメッセージ
          example: 入力内容に誤りがあります
        details:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          description: フィールドレベルのバリデーションエラー詳細（VALIDATION_ERRORの場合のみ）
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: エラーが発生したフィールド名
          example: userCode
        message:
          type: string
          description: フィールドのエラーメッセージ
          example: 必須項目です
...
    ToggleActiveRequest:
      type: object
      required: [isActive, version]
      properties:
        isActive:
          type: boolean
          description: falseで無効化、trueで有効化
        version:
          type: integer
          description: 楽観的ロックバージョン

    # --- Facility Master: Warehouse ---
...
    CreateBuildingRequest:
      type: object
      required: [warehouseId, buildingCode, buildingName]
      properties:
        warehouseId:
          type: integer
          format: int64
          description: 所属倉庫ID
        buildingCode:
          type: string
          maxLength: 10
          pattern: '^[A-Za-z0-9]+$'
          description: 棟コード（倉庫内一意・変更不可）
          example: A
        buildingName:
          type: string
          maxLength: 200
          description: 棟名称
          example: A棟
    UpdateBuildingRequest:
      type: object
      required: [buildingName, version]
      properties:
        buildingName:
          type: string
          maxLength: 200
          description: 棟名称
        version:
          type: integer
          description: 楽観的ロックバージョン
    BuildingDetail:
      type: object
      required: [id, buildingCode, buildingName, warehouseId, warehouseCode, isActive, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        buildingCode:
          type: string
        buildingName:
          type: string
        warehouseId:
          type: integer
          format: int64
        warehouseCode:
          type: string
        warehouseName:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BuildingPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BuildingListItem'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    BuildingListItem:
      type: object
      required: [id, buildingCode, buildingName, warehouseId, warehouseCode, isActive, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        buildingCode:
          type: string
        buildingName:
          type: string
        warehouseId:
          type: integer
          format: int64
        warehouseCode:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BuildingToggleResponse:
      type: object
      required: [id, buildingCode, buildingName, isActive, version, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        buildingCode:
          type: string
        buildingName:
          type: string
        isActive:
          type: boolean
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time

    # --- Facility Master: Area ---
...
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: リソースID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: ページ番号（0始まり）
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: ページサイズ
...
  responses:
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validationError:
              value:
                errorCode: VALIDATION_ERROR
                message: 入力内容に誤りがあります
                details:
                  - field: userCode
                    message: 必須項目です
    Unauthorized:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                errorCode: UNAUTHORIZED
                message: 認証が必要です
    Forbidden:
      description: 権限不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                errorCode: FORBIDDEN
                message: この操作を実行する権限がありません
    InvalidParameter:
      description: クエリパラメータ不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidParameter:
              value:
                errorCode: INVALID_PARAMETER
                message: クエリパラメータが不正です
    OptimisticLockConflict:
      description: 楽観的ロック競合
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            conflict:
              value:
                errorCode: OPTIMISTIC_LOCK_CONFLICT
                message: 他のユーザーによる更新が先行しました。最新データを取得してから再度お試しください
security:
  - cookieAuth: []